Add draft for HiBench SparkApp generation

+++ ansible/templates/hibench/start-deamons.sh
@@ -2,17 +2,17 @@
 # ansible is not sourcing .bashrc
 export JAVA_HOME=$(readlink -f $(which java) | sed "s:bin/java::")
 export PATH=$JAVA_HOME/bin:$PATH
-export PDSH_RCMD_TYPE=exec
+export PDSH_RCMD_TYPE=ssh
 
-# stop services which might are already running
-{{ hadoop_dir }}/sbin/stop-all.sh
-
-# hadoop
-sudo rm -rf /tmp/hadoop-{{ linux_username }}/*
-yes | {{ hadoop_dir }}/bin/hdfs namenode -format
-{{ hadoop_dir }}/sbin/start-dfs.sh
-{{ hadoop_dir }}/bin/hdfs dfs -mkdir /user
-{{ hadoop_dir }}/bin/hdfs dfs -mkdir /user/{{ linux_username }}
+# hdfs
+hdfs_address={{ hdfs_address }}
+# only run if HDFS is running locally (and not in k8s cluster)
+if [[ $hdfs_address == *"localhost"* ]]; then
+  sudo rm -rf /tmp/hadoop-{{ linux_username }}/*
+  yes | {{ hadoop_dir }}/bin/hdfs namenode -format
+  {{ hadoop_dir }}/sbin/start-dfs.sh
+fi
 
 # yarn
+{{ hadoop_dir }}/sbin/stop-yarn.sh
 {{ hadoop_dir }}/sbin/start-yarn.sh
